msm: kgsl: Remove POPP
POPP constantly attempts to lower the GPU's frequency behind the
governor's back in order to save power; however, the GPU governor in use
(msm-adreno-tz) is very good at determining the GPU's load and selecting
an appropriate frequency to run the GPU at.

POPP was created long ago, perhaps when msm-adreno-tz didn't exist or
didn't work so well, so it is clearly deprecated. Remove it.

Conflicts:
	drivers/gpu/msm/kgsl_gmu.c
	drivers/gpu/msm/kgsl_pwrctrl.c
	drivers/gpu/msm/kgsl_pwrscale.h

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[fixed conflicts]
Signed-off-by: Volodymyr Zhdanov <wight554@gmail.com>